# apps/web/lib/api/generated/openapi.merged.yaml
openapi: 3.1.0
info:
  title: Zimmamless V3 API
  version: 3.1.0
  description: |
    Receivables marketplace connecting Jordanian suppliers with banks.

    CROSS-CUTTING RULES — binding on every endpoint:

    1. AUTH. Bearer JWT from Supabase Auth. Every request must also send
       `X-Organization-Id` naming the active organization context.
       Requests without it, or naming an org the user does not belong to,
       return 403.

    2. CONFIDENTIALITY. `minimumAcceptableAmount` MUST NEVER appear in any
       bank-facing response, including validation errors. A bank MUST NEVER
       receive another bank's identity, amount, fees, conditions, or the
       count of competing offers.

    3. MONEY. All monetary values are strings in decimal form with exactly
       three decimal places, e.g. "1250.000". Never numbers — JSON floats
       lose precision. Currency is always "JOD" in V3.

    4. IDEMPOTENCY. All POST endpoints that move money or change financial
       state require an `Idempotency-Key` header.

    5. LANGUAGE. `Accept-Language: en | ar`. Default is `en` (never inferred
       from locale). Affects messages and generated documents only.

    6. ERRORS. Every error uses the Error schema. `code` is stable and
       machine-readable; `message` is localized and for humans.
servers:
  - url: https://api.zimmamless.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: In-platform notifications for the current user in the active context
      parameters:
        - name: unread
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/NotificationItem'
                  unreadCount:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /notifications/{id}/read:
    post:
      tags:
        - Notifications
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Marked read
  /notifications/{id}/manual-call:
    post:
      tags:
        - Notifications
      summary: Record a call an operator actually made (platform staff only)
      description: |
        ZM-NOT-007 requires the manual call record with its recording user and outcome. `notifications.manual_call_notes` and `manual_call_by` exist in the frozen schema and nothing could write them. Deliberately separate from `/read`: a recipient opening their inbox and an operator attesting to a phone conversation are different claims by different people. The previous notes are retained in the audit entry — the column holds one value, so overwriting a colleague's account of a conversation with no trace would be a hard delete of evidence (INV-7).
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - notes
              properties:
                notes:
                  type: string
                  minLength: 1
                  maxLength: 4000
      responses:
        '200':
          description: Recorded
        '404':
          description: No MANUAL_CALL notification with that id
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    NotificationItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        templateKey:
          type: string
        subject:
          type: string
        body:
          type: string
        transactionId:
          type: string
          format: uuid
        read:
          type: boolean
        queuedAt:
          type: string
          format: date-time